Bug 1331828 - Fix ffi inconsistency of wr_readback_into_buffer() r=gfx?

+++ gfx/webrender_bindings/webrender_ffi.h
@@ -369,7 +369,7 @@ wr_dp_push_iframe(WrState* wrState, WrRect bounds, WrRect clip, WrPipelineId lay
 // and also free it at the proper time.
 WR_INLINE const uint8_t*
 wr_readback_into_buffer(WrWindowState* wrWindow, uint32_t width, uint32_t height,
-                        uint8_t* dst_buffer, uint32_t buffer_length)
+                        uint8_t* dst_buffer, size_t buffer_length)
 WR_FUNC;
 
 // TODO: Remove.
